Make function signature immutable by default
Originally, `func_types` was mutable and could be shared between the same class across multiple threads.

However, if someone modified `func_types`, it would disrupt the binding with the C ABI.

Therefore, making `func_types` immutable enhances its robustness.

There are two key changes:

- Tuples were converted to `namedtuple`. This gives the tuple meaning, makes it easier for newcomers to debug, and is generally a good practice.
- The dictionary was made immutable using `MappingProxyType`. If someone wants to change the dictionary, they must recompile the code.

The term `__fs` stands for **function signature**. The double underscore (`__`) indicates that users should not rely on any method or variable with two underscores, as it is private to us. This is considered good practice.

Co-authored-by: Duc Nguyen <ductri.nguyen@sandboxaq.com>
GitOrigin-RevId: 27a60e15ede43f6fb1c4dc9f6a6f1cd191e6eeac