QST-3: Add listener API
This PR adds a new set of helper functions aimed at giving the user an easy way
to accept incomming connections. This particular PR instantiates a TCP listener,
however it is not limited to only TCP (Turbo Transport, when?). The
Sandwich library will own the listener and any IO objects spawned by the listener.

At this time there is no way for a user to define their own sandwich listener object
and have it work with the sandwich_listener_* functions.

Co-authored-by: jgoertzen-sb jason.goertzen@sandboxquantum.com <>
GitOrigin-RevId: 9ff0ba60407de933a1c410cf6079245855fc0397