Refactor IOErrors.
This commit refactors the way we handle IO errors.

Currently, `crate::IO` methods return a value or an [`std::io::Error`].
We had a wrapper around `pb::IOError` that was unused.

This commit defines a simple trait called `IntoIOError` that converts a type
to a `pb::IOError` (`pb::IOError` are used in FFI).

This commit also combines types and definitions related to IO in FFI into a new module
called `io`, in order to have the same module layout in FFI as in the core library.

[`std::io::Error`]: https://doc.rust-lang.org/std/io/struct.Error.html

GitOrigin-RevId: e3e2e8c2958ea37f07c84eb1b04f4f75c76f13e7