Add `flush` method to Sandwich IO.
Since we're going to migrate from our custom IO trait to the generic trait
from Rust' std (`std::io::{Read,Write}`), we need a flush method in addition
to the read and write methods.

This commit implements this flush method. Note that it's optional.

The flush method is now called by our backend with requested.

GitOrigin-RevId: 45ccadac60e947d6a8650b7d0edd44857a2b6ab7